requirements.txt
discord.py
Pillow
torch
diffusers
transformers
